# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "KAVACH API",
        "description": "Risk, Fraud and Regulatory Intelligence Copilot for Indian Banking",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/why-not/{txn_id}": {
            "get": {
                "tags": [
                    "Analysis"
                ],
                "summary": "Why Not",
                "description": "Explain why a transaction didn't trigger an alert",
                "operationId": "why_not_api_why_not__txn_id__get",
                "parameters": [
                    {
                        "name": "txn_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Txn Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WhyNotResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "WhyNotResponse": {
                "properties": {
                    "txn_id": {
                        "type": "string",
                        "title": "Txn Id"
                    },
                    "explanation": {
                        "type": "string",
                        "title": "Explanation"
                    },
                    "rules_checked": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Rules Checked"
                    },
                    "recommendation": {
                        "type": "string",
                        "title": "Recommendation"
                    }
                },
                "type": "object",
                "required": [
                    "txn_id",
                    "explanation",
                    "rules_checked",
                    "recommendation"
                ],
                "title": "WhyNotResponse",
                "description": "Why-not response model"
            }
        }
    }
}